MAINT/DEL: Delete old abc conv runscript, refactor abc heterogeneity model. Delete old heterogeneity model.
